aria2: do not use execvp on NT
It appears that a process cannot be replaced with execvp on Windows
NT (?); when execvp is called, a new process is spawned while the
original process simply completes it, and when you run the command in
console, you regain control over the console as soon as execvp is
called, which is apparently undesirable.

Also, aria2c is no longer called when there's nothing to download.